move page not found component to report as it is the only place using it and the components dont actually work without predefined asset (its not reusable so not suitable for react-components)
