--- packages/react-components/lib/error-pages/index.ts
@@ -1,2 +1,0 @@
-export * from './page-not-found';
-export * from './unauthorized';
