+++ packages/react-components/lib/index.ts
@@ -3,7 +3,6 @@ import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
 
-export * from './alert-dialog';
 export * from './color-manager';
 export * from './commands';
 export * from './confirmation-dialog';
@@ -11,7 +10,6 @@ export * from './date-time-picker';
 export * from './dispensers';
 export * from './doors';
 export * from './error-overlay';
-export * from './error-pages';
 export * from './error-snackbar';
 export * from './form-inputs';
 export * from './geometry-utils';
